# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Telegram Bot API server",
        "title": "Telegram Bot API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/telegram/send": {
            "post": {
                "description": "Sends a message via Telegram Bot API to the specified chat ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Telegram"
                ],
                "summary": "Send message to Telegram user",
                "parameters": [
                    {
                        "description": "Message payload",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/telegramInterface.TelegramSendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message sent",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "telegramInterface.TelegramSendRequest": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}